🚚 Move and Remove Files: 불필요한 폴더 삭제 및 code style 수정을 위한 파일 분리
