+++ src/shared/components/index.ts
@@ -1,5 +1,5 @@
 //Barrel File
-export { Header } from "./Header";
-export { NewsCard } from "./NewsCard";
-export { Badge } from "./Badge";
-export { Carousel } from "./Carousel";
+export * from "./Header";
+export * from "./NewsCard";
+export * from "./Badge";
+export * from "./Carousel";
