--- src/shared/hooks/useIsMounted.tsx
@@ -1,13 +1,0 @@
-import { useEffect, useState } from "react";
-
-const useIsMounted = () => {
-  const [mounted, setMounted] = useState(false);
-
-  useEffect(function setMouteState() {
-    setMounted(true);
-  }, []);
-
-  return mounted;
-};
-
-export default useIsMounted;
