--- src/shared/models/test.tsx
@@ -1,1 +1,0 @@
-// 비즈니스 로직 담는 model 파일
