--- src/shared/stores/mockZustandStore.tsx
@@ -1,16 +1,0 @@
-// 상태 관리 관련 메소드
-
-// import { StoreApi } from "zustand";
-// import {
-//   ProblemModuleStore,
-//   useProblemModuleStore,
-// } from "./problemModuleStore";
-
-// const mockStore = <Hook, State>(hook: StoreApi<Hook>, state: State) => {
-//   const initStore = hook.getState();
-//   hook.setState({ ...initStore, ...state }, true);
-// };
-
-// export const mockProblemModuleStore = (state: Partial<ProblemModuleStore>) => {
-//   mockStore(useProblemModuleStore, state);
-// };
