+++ src/shared/utils/index.ts
@@ -1,17 +1,1 @@
-const formatKoreanDate = (date: Date) => {
-  return new Intl.DateTimeFormat("ko-KR", {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  }).format(date);
-};
-
-const formatDateToYYYYMMDD = (date: Date): string => {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-
-  return `${year}-${month}-${day}`;
-};
-
-export { formatKoreanDate, formatDateToYYYYMMDD };
+export * from "./date";
